+++ evolver/circle_mutator.go
@@ -9,11 +9,11 @@ import (
 
 type CircleMutator struct {
 	config      *Config
-	imageWidth  float64
-	imageHeight float64
+	imageWidth  float32
+	imageHeight float32
 }
 
-func NewCircleMutator(config *Config, imageWidth float64, imageHeight float64) *CircleMutator {
+func NewCircleMutator(config *Config, imageWidth float32, imageHeight float32) *CircleMutator {
 	mut := new(CircleMutator)
 	mut.config = config
 	mut.imageWidth = imageWidth
@@ -60,20 +60,20 @@ func (mut *CircleMutator) mutateColor(circle *Circle) {
 
 func (mut *CircleMutator) mutateHue(circle *Circle) {
 	hue, sat, lightness := MakeColor(circle.Color).Hsl()
-	newHue := mut.mutateValue(0, 360, mut.config.MinHueMutation, mut.config.MaxHueMutation, hue)
-	circle.Color = LoadColor(SaveColor(colorful.Hsl(newHue, sat, lightness)))
+	newHue := mut.mutateValue(0, 360, mut.config.MinHueMutation, mut.config.MaxHueMutation, float32(hue))
+	circle.Color = LoadColor(SaveColor(colorful.Hsl(float64(newHue), sat, lightness)))
 }
 
 func (mut *CircleMutator) mutateSaturation(circle *Circle) {
 	hue, sat, lightness := MakeColor(circle.Color).Hsl()
-	newSat := mut.mutateValue(0, 1, mut.config.MinSaturationMutation, mut.config.MaxSaturationMutation, sat)
-	circle.Color = LoadColor(SaveColor(colorful.Hsl(hue, newSat, lightness)))
+	newSat := mut.mutateValue(0, 1, mut.config.MinSaturationMutation, mut.config.MaxSaturationMutation, float32(sat))
+	circle.Color = LoadColor(SaveColor(colorful.Hsl(hue, float64(newSat), lightness)))
 }
 
 func (mut *CircleMutator) mutateLightness(circle *Circle) {
 	hue, sat, lightness := MakeColor(circle.Color).Hsl()
-	newLightness := mut.mutateValue(0, 1, mut.config.MinValueMutation, mut.config.MaxValueMutation, lightness)
-	circle.Color = LoadColor(SaveColor(colorful.Hsl(hue, sat, newLightness)))
+	newLightness := mut.mutateValue(0, 1, mut.config.MinValueMutation, mut.config.MaxValueMutation, float32(lightness))
+	circle.Color = LoadColor(SaveColor(colorful.Hsl(hue, sat, float64(newLightness))))
 }
 
 // Mutate Brush Size
@@ -96,20 +96,20 @@ func (mut *CircleMutator) mutateCoordinates(circle *Circle) {
 // Swap Instructions
 func (mut *CircleMutator) RandomInstruction() Instruction {
 	return &Circle{
-		X: rand.Float64() * mut.imageWidth,
-		Y: rand.Float64() * mut.imageHeight,
+		X: rand.Float32() * mut.imageWidth,
+		Y: rand.Float32() * mut.imageHeight,
 		Color: &color.RGBA{
 			A: 255,
 			G: uint8(rand.Int31n(255)),
 			B: uint8(rand.Int31n(255)),
 			R: uint8(rand.Int31n(255)),
 		},
-		Radius: rand.Float64()*(mut.config.MaxCircleRadius-1) + 1,
+		Radius: rand.Float32()*(mut.config.MaxCircleRadius-1) + 1,
 	}
 }
 
-func (mut *CircleMutator) mutateValue(min float64, max float64, minDelta float64, maxDelta float64, value float64) float64 {
-	amt := rand.Float64()*(maxDelta-minDelta) + minDelta
+func (mut *CircleMutator) mutateValue(min float32, max float32, minDelta float32, maxDelta float32, value float32) float32 {
+	amt := rand.Float32()*(maxDelta-minDelta) + minDelta
 	value = value + amt
 	// Make the new value wrap around at the inclusive boundaries
 	for value < min {
